Pictures of Progress
After two days of nonstop working, I was able to create a VEX robot in order to compete and showcase it to my school's club. I began this adventure because my school was hosting a club introduction to all of the incoming high schoolers and I really wanted to increase the number of members and teams in the club. I also had an offseason competition I wanted to attend so this would work for both. I began by creating a very quick drivetrain as well as experimenting with the catapult as I had never really constructed a catapult before. Overall, this was an amazing experience and I will definitely use the things I learned when creating a new robot for the next season.